db_builder/requirements.txt
APScheduler==3.10.4
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
loguru==0.7.2
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.1
parse-torrent-name==1.1.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.27
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
Werkzeug==3.0.1
